test(cors,csrf): consolidate added tests into existing test files
Move the coverage and security tests out of the standalone
coverage_test.go / security_test.go files and into the existing test
files, matching each test to the source file it exercises:

- CORS: all tests (cors.go helpers and security scenarios) -> cors_test.go
- CSRF: validateExtractorSecurity nil guard -> config_test.go;
  constant-time comparison helpers -> helpers_test.go;
  remaining coverage and security tests (extractor, storage/session
  managers, DeleteToken, cookie attributes, Sec-Fetch-Site, double
  submit, etc.) -> csrf_test.go

No test logic changed; only relocation plus the required imports.
